ecommerce/requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gspread==6.1.3
httplib2==0.22.0
idna==3.10
oauth2client==4.1.3
oauthlib==3.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
python-barcode==0.15.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.3
